backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.1
google-generativeai==0.3.2
